requirements.txt
django-analytical==1.0.0
